+++ tarpc/Cargo.toml
@@ -15,6 +15,7 @@ bincode = "^0.4.0"
 log = "^0.3.5"
 scoped-pool = "^0.1.5"
 serde = "^0.6.14"
+unix_socket = "^0.5.0"
 
 [dev-dependencies]
 lazy_static = "^0.1.15"
